Move Company details content and info header to admin's components
