+++ src/pages/Admin/components/Company/CompanyDetailContent/component.tsx
