+++ src/pages/Admin/components/Company/CompanyDetailContent/index.ts
