+++ src/pages/Admin/components/Company/CompanyDetailInfo/MainTitle/index.ts
