+++ src/pages/Admin/components/Company/CompanyDetailInfo/UserDetails/component.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from "react";
 import PersonOutlinedIcon from "@material-ui/icons/PersonOutlined";
 import EmailOutlinedIcon from "@material-ui/icons/EmailOutlined";
-import { TaskHeaderInfo } from "../../../../../../components/TaskHeaderInfo";
+import { TaskHeaderInfo } from "../../../TaskHeaderInfo";
 import { IUserDetailsProps } from "./interfaces";
 
 import styles from "./styles.module.scss";
