+++ src/pages/Admin/components/Company/CompanyDetailInfo/component.tsx
@@ -6,7 +6,7 @@ import { IUser } from "$interfaces/User";
 
 import { UserDetails } from "./UserDetails";
 import { MainTitle } from "./MainTitle";
-import { DetailInfo } from "../../../../../components/DetailInfo";
+import { DetailInfo } from "../../DetailInfo";
 
 export const CompanyDetailInfo: FunctionComponent<ICompanyDetailInfoProps> = ({
   setStatus,
