+++ src/pages/Admin/components/Company/CompanyDetailInfo/index.ts
